requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8

# Embeddings & ML
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

# Vector Database
pinecone-client==3.0.2

# LLM
groq==0.4.2

# Utilities
numpy==1.24.3
tqdm==4.66.1
typing-extensions==4.9.0